Move `alpha.api_client` to `alpha.internal.api_client`
